+++ tests/test_format.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 
-from bitformat import Format, Dtype, DtypeSingle, Bits, Field, Array, FieldType, Repeat, DtypeKind, Expression, While
+from bitformat import Format, Dtype, DtypeSingle, Bits, Field, Array, FieldType, Repeat, DtypeKind, Expression, While, \
+    MutableBits
 from hypothesis import given
 import pytest
 import hypothesis.strategies as st