+++ release_notes.md
@@ -1,11 +1,12 @@
 # Release Notes
 
-### Not yet released: version 0.8.
+### September 2025: version 0.8.
 
 * Allowing property setters for `MutableBits`. So for example you can now write `b.u = 34` to set the unsigned int
   interpretation of `b` to 34 (for whatever length `b` currently has).
 * Slice assignments / deletion are more capable, so you can set slices with a step not equal to 1, for example.
 * from_random() now takes a bytes object as the optional seed instead of an int.
+* Several fixes and internal changes to support the experimental version of bitstring.
 
 ### August 2025: version 0.7.